COmplete overhaul
