+++ addons/adminMenu/functions/fnc_uihook_safeStart.sqf
@@ -3,14 +3,12 @@
 params ["_newSetting"];
 TRACE_1("params",_this);
 
-if (_newSetting) then {
-    [] remoteExecCall [QEFUNC(safeStart,makeSafe), 2];
+[_newSetting] remoteExecCall [QEFUNC(safeStart,toggleSafeStart), 2];
 
-    private _debugMsg = format ["Turning on SafeStart"];
-    ["potato_adminMsg", [_debugMsg, profileName]] call CBA_fnc_globalEvent;
-} else {
-    missionNamespace setVariable [QEGVAR(safeStart,startTime_PV), -1, true];
-
-    private _debugMsg = format ["Turning off SafeStart"];
-    ["potato_adminMsg", [_debugMsg, profileName]] call CBA_fnc_globalEvent;
-};
+[
+    "potato_adminMsg",
+    [
+        ["Turning off SafeStart", "Turning on SafeStart"] select _newSetting,
+        profileName
+    ]
+] call CBA_fnc_globalEvent;
