+++ addons/safeStart/ACE_Settings.hpp
@@ -1,0 +1,14 @@
+class ACE_Settings {
+    class GVAR(showTimer) {
+        category = QUOTE(PREFIX);
+        displayName = "Potato: Show SafeStart Timer";
+        description = "Show the safe start timer";
+        typeName = "BOOL";
+        value = 1;
+        isClientSettable = 1;
+    };
+    class GVAR(enabled) {
+        typeName = "BOOL";
+        value = 0;
+    };
+};
