+++ addons/safeStart/XEH_postInit.sqf
@@ -1,18 +1,21 @@
 #include "script_component.hpp"
 
-["potato_safeStartOn", {
-    GVAR(safeStartEnabled) = true;
-    [] call FUNC(showTimer);
-}] call CBA_fnc_addEventHandler;
+if (isServer) then {
+    [
+        { ACEGVAR(common,settingsInitFinished) && {time > 0} },
+        { [GVAR(enabled)] call FUNC(toggleSafeStart); }
+    ] call CBA_fnc_waitUntilAndExecute;
+};
 
-[
-    {
-        ACEGVAR(common,settingsInitFinished) && {time > (_this select 0)}
-    }, {
-        if (GVAR(enabled)) then {
-            [] call FUNC(makeSafe);
-        } else {
-            GVAR(safeStartEnabled) = false;
-        };
-    }, [0]
-] call CBA_fnc_waitUntilAndExecute;
+if (didJip) then {
+    [
+        { ACEGVAR(common,settingsInitFinished) && {time > 1} },
+        {
+            if (missionNamespace getVariable [QGVAR(startTime_PV), -1] == -1) then {
+                ["potato_safeStartOff"] call CBA_fnc_localEvent;
+            } else {
+                ["potato_safeStartOn"] call CBA_fnc_localEvent;
+            };
+        }
+    ] call CBA_fnc_waitUntilAndExecute;
+};
