+++ addons/safeStart/config.cpp
@@ -13,20 +13,6 @@ class CfgPatches {
     };
 };
 
+#include "ACE_Settings.hpp"
 #include "CfgEventHandlers.hpp"
 #include "gui_safeStart.hpp"
-
-class ACE_Settings {
-    class GVAR(showTimer) {
-        category = QUOTE(PREFIX);
-        displayName = "Potato: Show SafeStart Timer";
-        description = "Show the safe start timer";
-        typeName = "BOOL";
-        value = 1;
-        isClientSettable = 1;
-    };
-    class GVAR(enabled) {
-        typeName = "BOOL";
-        value = 0;
-    };
-};
