+++ addons/safeStart/functions/script_component.hpp
