+++ addons/safeStart/script_component.hpp
@@ -13,4 +13,7 @@
     #define DEBUG_SETTINGS DEBUG_SETTINGS_SAFESTART
 #endif
 
+#define TIMER_LAYER "SafeStartTimerRscLayer"
+#define START_LAYER "SafeStartStartingRscLayer"
+
 #include "\z\potato\addons\core\script_macros.hpp"
